requirements.txt
fastapi==0.95.0
line-bot-sdk==2.4.2
pydantic